Moved AIProduct* to ai/description package
