+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/products/ai/description/AIProductDescriptionBottomSheetFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.products
+package com.woocommerce.android.ui.products.ai.description
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -11,7 +11,7 @@ import com.woocommerce.android.R
 import com.woocommerce.android.extensions.copyToClipboard
 import com.woocommerce.android.extensions.navigateBackWithResult
 import com.woocommerce.android.ui.compose.theme.WooThemeWithBackground
-import com.woocommerce.android.ui.products.AIProductDescriptionViewModel.CopyDescriptionToClipboard
+import com.woocommerce.android.ui.products.ai.description.AIProductDescriptionViewModel.CopyDescriptionToClipboard
 import com.woocommerce.android.util.WooLog
 import com.woocommerce.android.viewmodel.MultiLiveEvent.Event.ExitWithResult
 import com.woocommerce.android.widgets.WCBottomSheetDialogFragment
